Rename `max` and `min` to avoid conflict

+++ data_structures/priority_queue.h
@@ -1,36 +1,34 @@
 #pragma once
 
-#include <cstddef>
-
 #include <utils.h> //expand: true
 
 template <typename T> class PriorityQueue {
 private:
   T *internal_arr;
-  size_t height, node_count, capacity;
-  void normalize_subtree(size_t root);
+  int height, node_count, capacity;
+  void normalize_subtree(int root);
 
 public:
-  PriorityQueue<T>(size_t height);
+  PriorityQueue<T>(int height);
   PriorityQueue<T>();
-  PriorityQueue<T>(const T *arr, size_t node_count);
+  PriorityQueue<T>(const T *arr, int node_count);
   ~PriorityQueue<T>();
   T *get_internal_array() const;
   bool empty() const;
-  size_t size() const;
+  int size() const;
   T minimum() const;
   void insert(T element);
   T remove();
 };
 
-template <typename T> void PriorityQueue<T>::normalize_subtree(size_t root) {
-  size_t current_index = root;
+template <typename T> void PriorityQueue<T>::normalize_subtree(int root) {
+  int current_index = root;
   while (true) {
-    size_t left_child_index = 2 * current_index,
-           right_child_index = 2 * current_index + 1;
+    int left_child_index = 2 * current_index,
+        right_child_index = 2 * current_index + 1;
     if (left_child_index > node_count && right_child_index > node_count)
       break;
-    size_t new_index = current_index;
+    int new_index = current_index;
     if (left_child_index <= node_count &&
         internal_arr[left_child_index] < internal_arr[new_index])
       new_index = left_child_index;
@@ -45,7 +43,7 @@ template <typename T> void PriorityQueue<T>::normalize_subtree(size_t root) {
 }
 
 template <typename T>
-PriorityQueue<T>::PriorityQueue(size_t height)
+PriorityQueue<T>::PriorityQueue(int height)
     : height(height), node_count(0), capacity(1 << height) {
   internal_arr = new T[capacity];
 }
@@ -53,7 +51,7 @@ PriorityQueue<T>::PriorityQueue(size_t height)
 template <typename T> PriorityQueue<T>::PriorityQueue() : PriorityQueue<T>(1) {}
 
 template <typename T>
-PriorityQueue<T>::PriorityQueue(const T *arr, size_t node_count)
+PriorityQueue<T>::PriorityQueue(const T *arr, int node_count)
     : node_count(node_count) {
   if (node_count == 0) {
     height = 1;
@@ -61,12 +59,12 @@ PriorityQueue<T>::PriorityQueue(const T *arr, size_t node_count)
     internal_arr = new T[capacity];
     return;
   }
-  height = log2_floor<size_t>(node_count) + 1;
+  height = log2_floor<int>(node_count) + 1;
   capacity = 1 << height;
   internal_arr = new T[capacity];
-  for (size_t i = 1; i <= node_count; i++)
+  for (int i = 1; i <= node_count; i++)
     internal_arr[i] = arr[i - 1];
-  for (size_t k = node_count / 2; k >= 1; k--)
+  for (int k = node_count / 2; k >= 1; k--)
     normalize_subtree(k);
 }
 
@@ -82,9 +80,7 @@ template <typename T> bool PriorityQueue<T>::empty() const {
   return node_count == 0;
 }
 
-template <typename T> size_t PriorityQueue<T>::size() const {
-  return node_count;
-}
+template <typename T> int PriorityQueue<T>::size() const { return node_count; }
 
 template <typename T> T PriorityQueue<T>::minimum() const {
   return internal_arr[1];
@@ -100,9 +96,9 @@ template <typename T> void PriorityQueue<T>::insert(T element) {
     internal_arr = new_arr;
   }
   internal_arr[node_count + 1] = element;
-  size_t current_index = ++node_count;
+  int current_index = ++node_count;
   while (current_index != 1) {
-    size_t parent_index = current_index / 2;
+    int parent_index = current_index / 2;
     if (internal_arr[current_index] >= internal_arr[parent_index])
       break;
     swap(internal_arr[current_index], internal_arr[parent_index]);
